Build yearly mart from the fact table and cast date parts to int

+++ dbt_climate/models/marts/dim_date.sql
@@ -5,9 +5,9 @@ with timeline as (
 
 select
     date_day,
-    date_part(year,  date_day) as year,
-    date_part(month, date_day) as month,
-    date_part(day,   date_day) as day,
+    date_part(year,  date_day)::int as year,
+    date_part(month, date_day)::int as month,
+    date_part(day,   date_day)::int as day,
     case when date_part(month, date_day) between 3 and 5 then 'spring'
          when date_part(month, date_day) between 6 and 8 then 'summer'
          when date_part(month, date_day) between 9 and 11 then 'fall'
